requirements.txt
# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0

# MongoDB
pymongo>=4.5.0
dnspython>=2.4.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
joblib>=1.3.0

# Visualization & Dashboard
streamlit>=1.28.0
plotly>=5.17.0
matplotlib>=3.7.0

# Utilities
python-dotenv>=1.0.0
schedule>=1.2.0
python-dateutil>=2.8.2

# Development
jupyter>=1.0.0
ipython>=8.15.0
black>=23.9.1
pytest>=7.4.3
# Core
pymongo==4.6.0
python-dotenv==1.0.0
pandas==2.0.3
numpy==1.24.3

# ML
scikit-learn==1.3.0
xgboost==1.7.6
prophet==1.1.5
tensorflow==2.13.0

# Visualization
plotly==5.17.0
streamlit==1.28.0

# Monitoring
schedule==1.2.0
loguru==0.7.2

# Testing
pytest==7.4.2
